requirements.txt
Flask_WTF==0.14.3
Flask==1.1.2
Flask_SQLAlchemy==2.4.3
WTForms==2.3.1
Werkzeug==1.0.1
Pillow==7.2.0
